+++ src/command_modules/azure-cli-billing/azure/cli/command_modules/billing/__init__.py
@@ -12,9 +12,10 @@ class BillingCommandsLoader(AzCommandsLoader):
 
     def __init__(self, cli_ctx=None):
         from azure.cli.core.commands import CliCommandType
+        from azure.cli.core.profiles import ResourceType
         billing_custom = CliCommandType(operations_tmpl='azure.cli.command_modules.billing.custom#{}')
         super(BillingCommandsLoader, self).__init__(cli_ctx=cli_ctx, custom_command_type=billing_custom,
-                                                    min_profile="2017-03-10-profile")
+                                                    resource_type=ResourceType.MGMT_BILLING)
 
     def load_command_table(self, args):
         from azure.cli.command_modules.billing.commands import load_command_table