+++ src/command_modules/azure-cli-cognitiveservices/azure/cli/command_modules/cognitiveservices/__init__.py
@@ -14,12 +14,13 @@ class CognitiveServicesCommandsLoader(AzCommandsLoader):
 
     def __init__(self, cli_ctx=None):
         from azure.cli.core.commands import CliCommandType
+        from azure.cli.core.profiles import ResourceType
         custom_type = CliCommandType(
             operations_tmpl='azure.cli.command_modules.cognitiveservices.custom#{}',
             client_factory=cf_accounts)
         super(CognitiveServicesCommandsLoader, self).__init__(cli_ctx=cli_ctx,
                                                               custom_command_type=custom_type,
-                                                              min_profile='2017-03-10-profile')
+                                                              resource_type=ResourceType.MGMT_COGNITIVESERVICES)
 
     def load_command_table(self, args):
         from azure.cli.command_modules.cognitiveservices.commands import load_command_table