+++ src/command_modules/azure-cli-dla/azure/cli/command_modules/dla/__init__.py
@@ -13,9 +13,10 @@ class DataLakeAnalyticsCommandsLoader(AzCommandsLoader):
 
     def __init__(self, cli_ctx=None):
         from azure.cli.core.commands import CliCommandType
+        from azure.cli.core.profiles import ResourceType
         dla_custom = CliCommandType(operations_tmpl='azure.cli.command_modules.dla.custom#{}')
         super(DataLakeAnalyticsCommandsLoader, self).__init__(cli_ctx=cli_ctx,
-                                                              min_profile='2017-03-10-profile',
+                                                              resource_type=ResourceType.MGMT_DATALAKE_ANALYTICS,
                                                               custom_command_type=dla_custom)
 
     def load_command_table(self, args):