+++ src/command_modules/azure-cli-maps/azure/cli/command_modules/maps/__init__.py
@@ -13,12 +13,13 @@ class MapsCommandsLoader(AzCommandsLoader):
 
     def __init__(self, cli_ctx=None):
         from azure.cli.core.commands import CliCommandType
+        from azure.cli.core.profiles import ResourceType
         custom_type = CliCommandType(
             operations_tmpl='azure.cli.command_modules.maps.custom#{}',
             client_factory=cf_accounts)
         super(MapsCommandsLoader, self).__init__(cli_ctx=cli_ctx,
                                                  custom_command_type=custom_type,
-                                                 min_profile='2017-03-10-profile')
+                                                 resource_type=ResourceType.MGMT_MAPS)
 
     def load_command_table(self, args):
         from azure.cli.command_modules.maps.commands import load_command_table