+++ src/command_modules/azure-cli-search/azure/cli/command_modules/search/__init__.py
@@ -12,10 +12,11 @@ class AzureSearchCommandsLoader(AzCommandsLoader):
 
     def __init__(self, cli_ctx=None):
         from azure.cli.core.commands import CliCommandType
+        from azure.cli.core.profiles import ResourceType
         azure_search_custom = CliCommandType(operations_tmpl='azure.cli.command_modules.search.custom#{}')
 
         super(AzureSearchCommandsLoader, self).__init__(cli_ctx=cli_ctx,
-                                                        min_profile='2017-03-10-profile',
+                                                        resource_type=ResourceType.MGMT_SEARCH,
                                                         custom_command_type=azure_search_custom)
 
     def load_command_table(self, args):